NCBC-2738: Improve perf adding tags for K/V spans
Motivation
----------
The IPEndPoints and connection IDs are constant for any given connection
but are being formatted as a string for every K/V operation.

Modifications
-------------
Format the tag values a string once for each connection as it is
constructed and reuse them for each new operation span.

Results
-------
Reduced CPU and heap allocations formatting strings for each key/value
operation.

Change-Id: Ic8e4770dd722ead78d4ed5278dd58be5a47c3959
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/141561
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: <richard.ponton@couchbase.com>