NCBC-2742: Reduce debug logging heap allocs for K/V ops
Motivation
----------
On the critical path for K/V operations our debug logs are adding a lot
of object[] heap allocations for the log parameters.

Modifications
-------------
Check if debug logging is enabled before the call to LogDebug so that
it isn't necessary to allocate an object[] on the heap to pass in as a
params array.

Results
-------
Normally, the cost of the log checks isn't very valuable, especially
since it rechecks again when you call LogDebug. However, for high
volume K/V transactions with Debug disabled (a much more common case
than enabled) the reduction in garbage collection is significant.

Change-Id: I720052210dd8536a8ccccf9892fed247dcefccb3
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/141709
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>