NCBC-2745: Reduce op completion lambda heap allocations
Motivation
----------
The current approach of passing a callback lambda from the operation to
the connection adds some additional heap allocations that can be avoided
since the lambda is always the same method.

Modifications
-------------
Expose OperationBase.HandleOperationCompleted on the IOperation
interface, and change IConnection.SendAsync to forward the entire
IOperation instead of an Action that referes to that method. Forward
this down to the callback points in AsyncState.

Move the read of the operation status from AsyncState to
HandleOperationCompleted. Since the method now has a single parameter,
it's possible to pass this to Task.Factory.StartNew as a state
parameter. This avoids a heap allocation to capture local variables.

Results
-------
Further reduction in heap allocations on K/V operations.

Change-Id: I9bb8993c6a1d3c50b9f70514deb157a63286a822
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/141758
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>