NCBC-2741: Handle send queue full gracefully after connection cleanup
Motivation
----------
Operations requeued after connection cleanup when the send queue is full are dropped on the floor.

Modifications
-------------
Set SendQueueFullException on the operation

Results
-------
Anytime the send queue is full the operation will be retried with
backoff until the timeout is reached, rather than being dropped.

If the retries still fail a meaningful SendQueueFullException is
returned to the consumer.

Change-Id: If2a4c7e011b8a9013f87fbea9b7cf09afcb5970f
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/141712
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Brant Burnett <bburnett@centeredgesoftware.com>