NCBC-2753: Allow BucketBase.RetryAsync to be inlined
Motivation
----------
BucketBase.RetryAsync is virtual but is never overridden except in unit
tests.

Modifications
-------------
Remove the virtual keyword from the method declaration. Adjust tests to
account for this limitation.

Results
-------
The JIT compiler may now choose to inline this method, and is very
likely to do so given its simplicity. This can provide a small
performance improvement.

Change-Id: Ib9466c6dd229a8b61e59b0facb4e1e3a5f8120ab
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/142020
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>