NCBC-2755: Use ThrowHelper to improve inlining
Motivation
----------
Many more methods could be inlined to improve performance by not
directly throwing exceptions.

Modifications
-------------
Create ThrowHelper as an internal static class to contain helpers that
throw common exceptions.

Replace throw statements with calls to ThrowHelper along a lot of hot
paths which are likely to be inlined.

Results
-------
The JIT compiler has a lot more options to inline a variety of methods
and therefore improve performance. Heuristics may not cause them to all
be inlined, but many of them will be.

Change-Id: I2afb42ac8b5ee47b27c218f800fe15b2d4078bf9
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/142034
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>