NCBC-2769: Avoid heap allocations for default K/V options
Motivation
----------
For the common scenario where a K/V operation is passed null options,
meaning to use the default, reduce heap allocations by using a shared
and immutable default instance of the options.

Modifications
-------------
Add an internal, static, and readonly Default property to all option
types (except for two cases which already had a public property). Use
this on all K/V operations instead of a new set of options when the
provided options are null.

Make the setters on all internal option properties private to ensure
that any internal usage goes through the mutation methods. Add a
Debug.Assert to the mutation methods to ensure we don't accidentally
attempt to mutate the default instance.

Change GetOptions.ProjectList to a ReadOnlyCollection to ensure it is
only mutated via the mutation method as well.

For GetAllReplicas and GetAnyReplica, approach them differently since
they already had a public default. Instead of preventing mutation of
the default, recognize the default and return a new object if an
attempt is made to mutate the default.

Results
-------
With the exception of GetAllReplicas and GetAnyReplica, there is no CPU
penalty for this change since Debug.Assert is optimized out of release
builds. However, we should catch any incorrect usage in our integration
and unit tests.

Heap allocations for the default case will be reduced by one allocation
per operation, two allocations in the case of Get ops with projections.

The public API surface is mostly unaffected. The only exception is an
API consumer missuing GetAllReplicas.Default or GetAnyReplica.Default.
If they were assuming mutation calls always mutate in place and not
using the return value, then they will not get the desired options.

```cs
// This misuse no longer works
var options = GetAllReplicas.Default;
options.CancellationToken(token);
// at this line, options will still be the default

// Correct usage:
var options = GetAllReplicas.Default.CancellationToken(token);
```

Change-Id: Ie378c9ad09552eff93054b3ac6d8db1167f5d9bc
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/142695
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: <richard.ponton@couchbase.com>