NCBC-2773: Simplify AsyncState handling of Opaque
Motivation
----------
Now that AsyncState keeps a pointer to the IOperation, rather than only
keeping a callback, we can get the opaque value directly from the
operation instead of extracting it from the buffer.

Modifications
-------------
Drop the opaque constructor parameter from AsyncState and change the
property to simply return the value from IOperation.Opaque. Drop the
extraction from the buffer in MultiplexingConnection.

Results
-------
Handling of opaque values on AsyncState is now much cleaner and easier
to understand, and should be slightly more performant as well.

Change-Id: Ie6eb42399e6b8fb57871e15c4b5197a1018e737a
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/142716
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>