NCBC-2775: Improve array handling performance throughout the SDK
Motivation
----------
There are a lot of small spots where arrays could be handled in a more
performant manner which may add up. Most of the hot path points are
already optimized, but we should get the other spots as well.

Modifications
-------------
Replaced a lot of spots that were allocating empty arrays with a call to
Array.Empty<T>.

When supported by the runtime, use stackalloc'd spans to generate
random numbers instead of allocating a byte array on the heap.

In ScramShaMechanism, do an in-place XOR on the byte array rather than
allocating a new byte array, since the original is not required.

In Hello and TrySetKeepAlives use MemoryMarshal.Cast to set the values
in the byte array more efficiently.

In VBucketKeyMapper, use memcpy to more efficiently build the array of
replicas, and use Array.Empty when there are no replicas.

Results
-------
Fewer heap allocations of empty arrays, using a shared copy since they
are inherently immutable. Also avoids heap allocations in several other
locations, as well as applying some CPU optimizations.

Change-Id: I389f515cfadaf10ca1caf7758f70de47fbd394b0
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/142720
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: <richard.ponton@couchbase.com>