NCBC-2776: Make SlicedMemoryOwner a structure
Motivation
----------
Reduce unnecessary key/value heap allocations.

Modifications
-------------
Rewrote SlicedMemoryOwner<T> to be a readonly structure with a default
that is empty. Since the structure is larger than the size of an IntPtr,
mark it as readonly and pass it to methods using `in` parameters (this
avoids the cost of making copies on the stack).

Rework AsyncState completion callbacks to reduce the heap allocations
required building actions, etc. Without this rewrite more allocations
would have been required to create a closure.

Drop the now unnecessary EmptyMemoryOwner class.

Fix a couple of unstable unit tests for channel connection pools.

Results
-------
There are at least 2 fewer heap allocations during the completion of
key/value operations. The Action<object> callback object in async state
is now static, there is no more SlicedMemoryOwner<byte> allocation
during the read of operations on the connection, and for GET operations
there is no second SlicedMemoryOwner<byte> allocation during the call
to OperationBase.ExtractBody.

Change-Id: Ib527bb4416e7dc84e2eca498a50ad5b9d63b8400
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/144346
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>