NCBC-2804: Allow doc get with expiry and non-JSON transcoder
Motivation
----------
When performing GetAsync and requesting Expiry using the options, an
exception occurs if you are using a non-JSON transcoder, such as the
RawBinaryTranscoder.

Modifications
-------------
Add an internal property to LookupInOptions that allows the transcoder
to be passed into lookup operations, and pass this in from GetAsync.
Detect the case where the whole document was requested (but not for
projection optimization) and run the data through the transcoder.

Add a static DefaultSerializer instance and made it the default for
JsonTranscoder.

Use the static DefaultSerializer for reading expiry in GetResult, since
ITypeTranscoder.Serializer may be null. Add responsible null checks at
other places.

Results
-------
Get operations now function with non-JSON transcoders, such as the
RawBinaryTranscoder, when you also request the expiry.

Such operations are now more performant with JSON transcoders, as an
intermediate step using dynamic JToken objects is eliminated.

A custom serializer won't cause unexpected behavior getting the expiry.

There are fewer DefaultSerializer instances allocated on the heap for
things like ClusterNode.GlobalTranscoder and the default DI transcoder.

Change-Id: I411ede7a10185abb02e3aca72abfd224797bc880
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/144432
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>