NCBC-2805: Use a cleanup loop for orphaned in-flight ops
Motivation
----------
Reduce heap allocations and CPU cycles related to cleaning up operations
which never receive a response from the server.

Modifications
-------------
Drop the Timer from AsyncState, and instead track in-flight time using a
Stopwatch. Drop the Cancel method, since it isn't actually necessary to
cancel the operation (that is handled by the operation timeout). Move
the more basic logic to the Dispose method.

Within the InFlightOperationSet, run a poll for operation which have
been in flight more than the timeout (currently 75 seconds), observing
them every 30 seconds. Drop the creation of the Timer and the related
handler.

Pass the missed CancellationToken to _stream.WriteAsync in
MultiplexingConnection.

Results
-------
Orphaned operations are removed from the in-flight operation set within
105 seconds, instead of the previous removal at precisely 75 seconds.
Since the purpose is merely to prevent memory leaks, this is sufficent.

Operations are still canceled via their operation timeout cancellation
tokens. When an in-flight op is determined to be orphaned, instead of
marking the whole operation as canceled instead only the AsyncState is
marked as done. This is used only for graceful connection shutdown
during connection pool scale down.

While it still causes a heap allocation, the stopwatch is a much more
lightweight alternative to timers. This benchmark compares creating and
disposing a Timer to creating and starting a Stopwatch:

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19042
Intel Core i7-9750H CPU 2.60GHz, 1 CPU, 12 logical and 6 physical cores
.NET Core SDK=5.0.102
  [Host]     : .NET Core 3.1.11 (CoreCLR 4.700.20.56602, CoreFX 4.700.20.56604), X64 RyuJIT
  DefaultJob : .NET Core 3.1.11 (CoreCLR 4.700.20.56602, CoreFX 4.700.20.56604), X64 RyuJIT

|    Method |      Mean |    Error |   StdDev | Ratio | Rank |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------- |----------:|---------:|---------:|------:|-----:|-------:|------:|------:|----------:|
|     Timer | 179.02 ns | 2.613 ns | 2.317 ns |  1.00 |    2 | 0.0331 |     - |     - |     208 B |
| Stopwatch |  22.17 ns | 0.221 ns | 0.196 ns |  0.12 |    1 | 0.0063 |     - |     - |      40 B |

Change-Id: Ie99a220583dfa859b3039ef6bf78f617e9b8f170
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/144574
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>