NCBC-2808: Use ValueTask and IValueTaskSource for OperationBase.Completed
Motivation
----------
Reduce heap allocations for each operation attempt related to the
TaskCompletionSource and Completed task. We can do this using ValueTask
because Completed is never awaited more than once.

Modifications
-------------
Drop the TaskCompletionSouce from OperationBase, and replace with a
ManualResetValueTaskSource. Return ValueTask for IOperation.Completed
instead of a Task. Use a _completed integer we can swap using
Interlocked.Exchange to emulate TrySetXXX behaviors from TCS.

Results
-------
Allocation of TaskCompleteSource and the Task returned by Completed has
been eliminated (approximately 96 bytes), in exchange for a slightly
larger OperationBase (approximately 50 bytes) that we're already
allocating on the heap. Additionally, there is no allocation on retries.

This is done in exchange for the limitation that the Completion task may
only be awaited once per retry. However, this was already the pattern.

The implementation of ManualResetValueTaskCore is being acquired from
the Microsoft.Bcl.AsyncInterfaces package for .NET Framework and .NET
Core 2.1. Due to being unable to access .NET runtime internals, it will
be slightly less performant on those frameworks. The type is built in to
.NET Core 3.0 and later.

Change-Id: I18171d0d86f13452f300e4029e0661b07425a93a
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/144673
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>