NCBC-2766: CreateScopeAsync not creating collections in ScopeSpec
Motivation
----------
CreateScopeAsync incorrectly has an overload that takes a ScopeSpec
which makes it appear that if the ScopeSpec contains CollectionSpecs,
they will be created undernieth the scope. This is incorrect, to create
collections CreateCollectionAsync will have to be called and the scope
name passed in to relate the two.

Modifications
-------------
 - Make CreateScopeAsync(ScopeSpec, options) obsolete.
 - Add CreateScopeAsync(string scopeName, ScopeOptions options).
 - Updated unit tests
 - Added overloads to extension methods and deprecated older methods

Result
------
The CollectionManager follows the spec as intended.

Change-Id: I8bf3a890ec8d5f6ae4018f6a7bfbc308b487fdcc
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/144674
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: <richard.ponton@couchbase.com>