NCBC-2784/NCBC-2767: getting a scope or collection after creation fails
Motivation
----------
If you use the Managment API to add a collection or scope and then
directly try to use it you will encounter a XxxNotFoundException. This
commit fixes that bug.

Modifications
-------------
 - Obsolete any sychronous scope or collection methods as we now may hit
 IO when they are fetched and do not exist in the local cache.
 - Add asynchronous methods for fetching scopes and collections to
 IBucket and implementations.
 - Add GetSid method for fetching the scope id directly by name.
 - Add unit tests
 - Fixup in old unit tests using the sychronous metheds.

Result
-------
Scopes and collections are now fetched and cached when directly
requested not by fetching the manifest. If you create a cluster with the
management API, you can begin using it as soon as its available from the
cluster.

Change-Id: Ide3abff60f97abce28b7718d9347450bb30609c6
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/144883
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: <richard.ponton@couchbase.com>