NCBC-2827: GET_CID and GET_SID do not correctly retry if Scope/Collection not found
Motivation
----------
GetSid and GetCid were defined on the IClusterNode interface and thus
were not able to be retried. If a collection was updated on the server
this would mean a Collection/ScopeOutdatedException would be thrown.
This fix moves the methods into the Bucket and Scope objects were the
operations can be retried on failure.

Modifications
-------------
 - Move GetSid to BucketBase
 - Move GetCid to Scope
 - Use Bucket.RetryAsync to fire off each operation
 - Update ScopeFactory and BucketFactory for tracing and
 IOperationConfiguration support
 - Make the CollectionNotFoundException and ScopeNotFoundException
 implement IRetriable to flag the failure as retriable
 - Add IOperation.RequiresVBucketId flag to ensure that operations that
 do not require VBucketId's have valid arguments (0)
 - Update logic in ClusterNode to assign VBucket id's only if required
 for an operation.
 - Update unit tests

Result
------
Scope and Collection identifier opperations are now retried if failed
with Scope/Collection not found.

Change-Id: I669edee7bf82072f0242b2f85ba14e561f77433b
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/146271
Reviewed-by: <richard.ponton@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>