NCBC-2828: Cleanup sub-doc operation public API surface
Motivation
----------
Reduce confusion for API consumers surrounding deprecated interfaces and
classes from SDK 2.x.

Modifications
-------------
Drop the use of MutateInBuilder and LookupInBuilder, which were only
used internally to collect OperationSpecs. Instead, just forward the
operation spec collection directly to MultiMutation and MultiLookup from
CouchbaseCollection. Move operation spec sorting in MultiLookup and
MultiMutation to the constructor instead of WriteBody to ensure that
the list is not resorted on retries, which would skew the OriginalIndex.

Delete MutateInBuilder, which was internal, and mark LookupInBuilder as
obsolete. Delete ISubdocInvoker which was internal. Mark other public
classes and interfaces related to Builders and Fragments as obsolete.

Change LookupInResult and MutateInResult to handle more specifically
typed OperationSpecs. Mark MutateInResult as obsolete so it can be made
internal in a future release.

Results
-------
The public API surface has pointless classes and interfaces clearly
marked as deprecated, and they can be removed in a future release.

There should also be a bit less processing and heap allocations around
the deprecated Builder classes when performing sub-doc operations.

Change-Id: I041bb7d1254b027adbbeae971d7fb549a38f1d43
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/146635
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>