NCBC-2842: Add Couchbase.Core.Exceptions.TimeoutException
Motivation
----------
There is a name collision between System.TimeoutException and
Couchbase.Core.Exceptions.TimeoutException, so previously the type was omitted
and UnambiguousTimeoutException and AmbiguousTimeoutException both derived
from CouchbaseException. This ticket to add
Couchbase.Core.Exceptions.TimeoutException and simply deal with the collision
by fully qualifying the namespace. This is for consisteny with other sdks and
alignment with the RFC.

Modifications
-------------
 - Add TimeoutException
 - Update AmbiguousTimeoutException and UnambiguousTimeoutExtension to
 use the new TimeoutException

Result
------
The exception hiearchy more closely resembles the RFC.

Change-Id: I6a79fbbe1910fb700521c1d84407a2443cbb46ba
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/148448
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>