NCBC-2849: Ensure FF Maps is only used after a NMVB
Motivation
----------
Reworks the FFMAP logic if a NMVB is encountered. Previously the FFMAP
was used if it existed, the error was NMVB or TransportError and the
revision of the cluster map returned in the body of a NMVB matched the
current cluster map revision. This fix changes the behavior to use the
FFMAP if a NMVB is encountered.

Modifications
-------------
 - Replace IKeyMapper.MapKey int parameter with boolean NMVB flag
 - Update all code using the MapKey method to pass in the flag
 - Remove TransportFailure as a retriable response
 - Add seam to allow the WasNmvb flag to be toggled for testing on
 operations
 - Update unit tests

Result
------
The .NET SDK will use FFMAPS whenever a NMVB is returned and the cluster
has a FFMAP. This aligns with Java behavior.

Change-Id: I00d03943b63197734e313f06fcb035f89a9c4356
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/150411
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: <richard.ponton@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>