NCBC-2858: Move collection id fetch into the operation call on the collection
Motivation
----------
Changes the behavior introduced in NCBC-2767; the scope and collection path
should not contain IO - it should be deferred until the operation method call
off of the collection itself.

Modifications
-------------
 - Reverted Obsolete attribute on async scope/collection calls as IO is
 no longer happening there.
 - Moved logic to fetch CID to just before the K/V operation in
 CouchbaseCollection
 - Removed the no longer dependencies on configurator and tracer from
 scope and collection.
 - Removed or fixed up any unit tests

Result
------
The CID fetch is deferred to the actual K/V operation call;
the SDK builds up an internal scope/collections structure, but
does not do any IO until the operation call.

Change-Id: I2642bb108ee20ce30f13f1ed4e0fcededc78d337
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/151337
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: <richard.ponton@couchbase.com>