NCBC-2839: send collection identifier as the body opposed to the key
Motivation
----------
A recent server change requires the collection identifier be sent as
part of the body of the GET_CID command as opposed to as the key due
to the size limitation in key size of 255 bytes.

Modifications
-------------
 - If the default collection/scope are used set the CID to zero (0)
 - Make the collection identifier be sent first as the part of the body
 and then if that fails as the key in case an older server is beng used.

Result
------
If supported the client will send the collection identifier as the body.
If this does not work, send as the key. Handle case where server doesn't
support collections at all with a default identifier of zero (0).

Change-Id: I55b4352b98ede3ba59b2c7dd3fed76a5156aa097
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/151430
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: <richard.ponton@couchbase.com>