NCBC-2868: ConfigHandler dead locks in K8 when delete pod is used
Motivation
----------
Fixes a bug where the ConfigHandler would go silent after deleting a k8s
pod and would not recover. The side effect was the SDK would not be
updated with the latest config and thus become stale, sending
UnAmbiguousTimeoutException when a key was mapped to the down node even
after failover.

Modifications
-------------
 - Add CancellationToken for the kv timeout duration in ClusterNode
 - Add additional logging in ConfigHandler

Result
------
The SDK will recover when the topology is changed in the cluster via a
delete pod and failover.

Change-Id: I40d3a8929558b6e293ac24ade089f08216786640
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/152828
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Brant Burnett <bburnett@centeredgesoftware.com>
Reviewed-by: <richard.ponton@couchbase.com>