NCBC-2876: Upserting to <7.0 clusters does not upsert the content
Motivation
----------
This affects 6.5 server and greater; it is caused by the NCBC-2839
ticket. In NCBC-2839 the scope.collection identifier was changed to be
passed as the body vs the key of the GET_CID command. The commit for
NCBC-2839 did this when collections were enabled; if they were not
enabled the default scope/collection Id was appended to the key which
caused data without a body to be stored in the server. In 7.0 a bug
introduced in NCBC-2839 caused similar behavior.

Modification
------------
 - First try to fetch the CID using the body, if this fails try again
 using the key approach - if this fails we are using a developer preview
 for collections and we do not send any CID.

Result
------
The SDK will now handle the three cases (>=7.0, 6.5-6.6 w/dev-preview
and without dev preview).

Change-Id: I3b7165076dac791588dfdffca73ea2fc12180419
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/153508
Reviewed-by: <richard.ponton@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>