NCBC-2574: Analytics Link management
Motivation:
Support Analytics Link CRUD for Cheshire Cat.

Modifications:
* Use separate records for Request/Response for each operation, with a repeating pattern to keep them as similar as possible.
* Add Builder pattern fallbacks for users who aren't using C# 9 or later.
* Unit tests and Integration tests

Remarks:
The general SDK RFC for dealing with the management API requires the request and response objects to be the same class.
I strongly disagree with the approach, since they have special rules for serialization/deserialization and don't flatten or unflatten the same in each case.
I've gotten as close as possible with the approach I took, where proxy classes are used for deserialization of the response internally.

Change-Id: I641432b640feb596eb036a08546e1d91ef6075fb
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/156386
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>