NCBC-2959: Allow sdk to send either the hostname or IP Address while as the target host
Motivation
----------
By default SDK3 sends the IP as the target host during TLS/SSL
authentication. This is exactly the opposite of SDK2 which sends either
hostname or the IP address depending upon what is returned in the server
configuration. This commit adds a flag, ForceIpAsTargetHost, which
allows the SDK to mimic SDK2 behavior and send either the hostname or IP
address. The default for this flag is true for backwards compatibility.

Modifications
-------------
 - Add ClusterOptions.ForceIpAsTargetHost which defaults to true.
 - Add additional parameter for HostEndpoint to IConnectionFactory.CreateAndConnectAsync
 - Update ConnectionFactory to either take the IP address and send it as
 the target host or the value returned in the server config, which could
 be IP or hostname.
 - Update unit tests

Result
------
Allows the default behavior in SDK3 of sending the IP to be changed to
the same behavior as in SDK2.

Change-Id: I9b6c721c1e8e3db06538011e89fa95ed12ea8f4b
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/160374
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>