NCBC-2966: Allow ILoggerFactory from DI to be overridden
Motivation
----------
It may be desired to provide a custom ILoggerFactory for Couchbase when
using Couchbase.Extensions.DependencyInjection. Currently that is not
possible, it always uses the global value from the DI container.

Modifications
-------------
Register an IConfigureOptions<ClusterOptions> in advance of the config
action which registers the ILoggerFactory on the options, and drop the
registration from the ClusterProvider.

Results
-------
Since the logger is set on ClusterOptions before the callback to the
configuration action, the configuration action may now override the
default behavior.

Change-Id: Ib3b5445dfd40b238db49194a37f2ad9289e22f7d
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/161769
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>