NCBC-2967: Rewrite OrphanReporter to avoid blocking calls
Motivation
----------
The current OrphanReporter is using Task.Factory.StartNew with
TaskCreationOptions.LongRunning to start the polling loop. This will
spin up a dedicated thread for the polling loop, as was clearly desired
since it's a long running process.

However, the method it is calling, DoWork, is also an async method. Once
the first "await Task.Delay" is reached, the created thread is done.
After the first delay is done, the remaining work will be scheduled back
on the ThreadPool. It will then block that thread when it reaches
_queue.TryTake (a BlockingCollection call).

Modifications
-------------
Rewrite OrphanReporter to use Channel<T> instead of
BlockingCollection<T> to avoid blocking altogether, and simply run on
the ThreadPool.

Results
-------
A ThreadPool thread won't become blocked during bootstrap, which may
reduce problems with warm up of a freshly started service.

Change-Id: I210ebf1e63e36d927c1439c27e35f398874d6c24
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/161890
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>