NCBC-2929: Improve client side error message when TLS is enforced on the server side
Motivation
----------
When connection encryption is enforced with 'strict' and non-TLS
connection will immediately be terminated. This fix adds additional
information to the logs that the encryption settings should be checked.

Modifications
-------------
 - Add ConnectException and message to check tls settings
 - Add more information to BucketNotFoundException.

Result
------
If a non-TLS connection is made while in `strict` mode the client will
fail with additional information to check the TLS settings.

Change-Id: I4fb97cf64580f1c94f34bb04fd30a3e1c4e51aa6
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/162017
Reviewed-by: <richard.ponton@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>