NCBC-2974: Don't release CidLock on timeout
Motivation
----------
It's possible that a timeout waiting for the CidLock may result in
incorrectly releasing the CidLock, despite the fact that the lock was
not taken by this thread.

Modifications
-------------
Only release CidLock if waitedSuccessfully == true, which indicates we
successfully acquired the lock.

Results
-------
Eliminates a corner case that could cause a race condition or exceptions
when the cluster is experiencing timeouts, for example during a
failover.

Change-Id: I4c3c8def62706876b729faca50e9e1ea168606fb
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/162354
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>