NCBC-2865:Improve bucket configuration handling with revEpoch
Motivation
----------
The server has added a new field, `revEpoch`, that allows the server to
provide higher level guidence with respect to current, correct
configurations. This commits checks if the revEpoch is present and if so
the SDK uses it otherwise the SDK falls back to the older rev field.

Modifications
-------------
 - Add BucketConfig.RevEpoch field
 - Add BucketConfigExtensions.IsNewer() method to determine if the
   config is newer than the current config.
 - Add unit tests

Result
------
If the revEpoch field is present, the SDK will use it to determine
configuration changes otherwise it will fall back to using the rev
field.

Change-Id: Ib1aaa40058064531fbc5be697137bb1b223125e6
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/164510
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>